Mqqt-client/requirements.txt
paho-mqtt
psycopg2-binarypip 


    # install -r requirements.txt
